api/portfolio/permissions.py
from rest_framework import permissions


class IsAuthenticatedOrReadOnly(permissions.BasePermission):
    """
    Custom permission to only allow authenticated users to modify objects,
    but allow anyone to read them.
    """

    def has_permission(self, request, view):
        # Allow GET requests without authentication
        if request.method in permissions.SAFE_METHODS:
            return True
        # Require authentication for other methods
        return request.user and request.user.is_authenticated
